src/subscription-transport-core/server has SubscriptionServer that doesn't depend on WebSocket
