+++ package.json
@@ -26,6 +26,7 @@
     "apollo-server": "^2.4.8",
     "apollo-server-lambda": "^2.4.8",
     "body-parser": "^1.18.3",
+    "cross-fetch": "^3.0.2",
     "fp-ts": "^1.17.0",
     "graphql": "^14.2.1",
     "grip": "^1.3.0",
