+++ src/subscriptions-transport-core/utils/is-string.ts
@@ -1,0 +1,3 @@
+export default function isString(value?: any): value is string {
+  return typeof value === "string";
+}
