+++ tslint.json
@@ -23,7 +23,9 @@
     "linterOptions": {
         "exclude": [
             "node_modules/**",
-            "src/subscriptions-transport-ws-over-http/**"
+            "src/subscriptions-transport-ws-over-http/**",
+            "src/subscriptions-transport-core/utils/**",
+            "src/subscriptions-transport-core/legacy/**"
         ]
     }
 }
